requirements.txt
arrow==1.3.0
beautifulsoup4==4.12.3
binaryornot==0.4.4
blinker==1.7.0
botcity-excel-plugin==0.1.7
botcity-framework-base==0.4.2
botcity-framework-core==1.1.2
botcity-framework-web==0.8.3
botcity-maestro-sdk==0.4.0
botcity-ms365-credentials-plugin==0.2.0
botcity-ms365-outlook-plugin==0.1.0
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.4.0
cookiecutter==2.6.0
cx_Oracle==8.3.0
distro==1.9.0
et-xmlfile==1.1.0
Flask==3.0.3
idna==3.6
importlib_metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
MouseInfo==0.1.3
msedge-selenium-tools==3.141.3
numpy==1.26.4
O365==2.0.34
oauthlib==3.2.2
opencv-python==4.9.0.80
openpyxl==3.1.2
packaging==24.0
pandas==2.1.4
pillow==10.3.0
psutil==5.9.8
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
Pygments==2.17.2
PyMsgBox==1.0.9
pynput==1.7.6
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
python-xlib==0.33
pytweening==1.2.0
pytz==2024.1
pywin32==306
pywinauto==0.6.8
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
selenium==3.141.0
setuptools==69.2.0
six==1.16.0
soupsieve==2.5
stringcase==1.2.0
text-unidecode==1.3
types-python-dateutil==2.9.0.20240316
tzdata==2024.1
tzlocal==5.2
urllib3==1.26.18
waitress==3.0.0
webdriver-manager==4.0.1
Werkzeug==3.0.2
zipp==3.18.1
